postfix-3.4-20190127

+++ postfix/src/tls/tls_mgr.c
@@ -441,7 +441,7 @@ int     main(int unused_ac, char **av)
 
 	    status = tls_mgr_lookup(argv->argv[1], argv->argv[2], buf);
 	    vstream_printf("status=%d session=%.*s\n",
-			   status, LEN(buf), STR(buf));
+			   status, (int) LEN(buf), STR(buf));
 	    vstring_free(buf);
 	} else if (COMMAND(argv, "update", 4)) {
 	    status = tls_mgr_update(argv->argv[1], argv->argv[2],
